badapple_generator/pyproject.toml
[project]
name = "ikea-led"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.0.0",
    "tqdm>=4.67.1",
]
